Update jscoverage.js
I like to move my lib folder to libArchive, use lib as the output directory for coverage and then when done delete lib and move libArchive back to lib. This enables me to not have to change my code with a lot of ugly process.env garbage and I don't have to .gitignore lib-cov or any of that stuff to keep it so that I don't get yelled at by jshint.